+++ packages/@expo/cli/src/start/server/BundlerDevServer.ts
@@ -452,6 +452,13 @@ export abstract class BundlerDevServer {
     );
   }
 
+  public getReactDevToolsUrl(): string {
+    return new URL(
+      '_expo/react-devtools',
+      this.getUrlCreator().constructUrl({ scheme: 'http' })
+    ).toString();
+  }
+
   protected async getPlatformManagerAsync(platform: keyof typeof PLATFORM_MANAGERS) {
     if (!this.platformManagers[platform]) {
       const Manager = PLATFORM_MANAGERS[platform]();